tested feature generation
